docs(benchmarks): make RESULTS.md a generated page, one row per model
The ledger was 291 lines, of which 174 were hand-written prose wrapped
around a 105-line generated block. It opened with three caveat
paragraphs, and the comparison a reader came for was 180 lines down.

RESULTS.md is now the generated block and nothing else, 81 lines. The
prose it carried is not deleted: measurements taken without a receipt,
the before/after studies against ferrox itself, and the sections that
predate the code they describe move to benchmarks/HISTORY.md, which is
where a generator cannot reach.

The table itself changes shape. It carried one row per (model, test),
so a ten-model host was twenty rows and a model's prefill and decode
sat ten rows apart. They are now one row: the two verdicts first,
because that is what a reader scans for, then the four raw tok/s
numbers. Rows sort by model name, since with both halves on one line
there is no single gap to sort on and the generated summary already
carries each host's range.

No number changed. Every gap value in the old file is present in the
new one, asserted by diffing the sorted multiset of gap cells across
RESULTS.md + HISTORY.md before and after.

tok/s now prints one decimal below 100 and none above. Whole tok/s
everywhere would have let a reader divide 149 by 151 and read 0.99
against a printed 0.98, because the gap column comes from the receipt
rather than from the two cells beside it.

Rendering moves to its own bench_render.rs: bench_suite.rs was 903
lines and this change needed to grow it, so the split came first.
903 becomes 432 + 678.

Nothing was re-benchmarked. `ferrox bench --render` reads the committed
receipts and rewrites the block; the receipts are untouched.
