+++ crates/ferrox-cli/src/bench_suite.rs
@@ -14,9 +14,9 @@ use anyhow::Context;
 use std::path::{Path, PathBuf};
 
 /// One `models[]` entry of `benchmarks/suite.json`.
-struct SuiteEntry {
-    id: String,
-    name: String,
+pub(crate) struct SuiteEntry {
+    pub(crate) id: String,
+    pub(crate) name: String,
     gguf: String,
     backends: Vec<String>,
     estimated_ram_gb: f64,
@@ -64,7 +64,7 @@ fn host_slug(label: &str) -> String {
 /// `host_spec.label`, which is a CPU string -- so a GTX 1080 row and an
 /// RTX 3080 row taken on the same Xeon sorted into one section, which
 /// is precisely the merge the grouping exists to prevent.
-fn host_identity(receipt: &serde_json::Value) -> String {
+pub(crate) fn host_identity(receipt: &serde_json::Value) -> String {
     let label = receipt
         .get("host_spec")
         .and_then(|h| h.get("label"))
@@ -80,11 +80,11 @@ fn suite_path(bench_dir: &Path) -> PathBuf {
     bench_dir.join("suite.json")
 }
 
-fn engine_receipt_dir(bench_dir: &Path) -> PathBuf {
+pub(crate) fn engine_receipt_dir(bench_dir: &Path) -> PathBuf {
     bench_dir.join("receipts").join("engine")
 }
 
-fn load_suite(bench_dir: &Path) -> anyhow::Result<Vec<SuiteEntry>> {
+pub(crate) fn load_suite(bench_dir: &Path) -> anyhow::Result<Vec<SuiteEntry>> {
     let path = suite_path(bench_dir);
     let text = std::fs::read_to_string(&path)
         .with_context(|| format!("reading suite at {}", path.display()))?;
@@ -295,383 +295,12 @@ pub fn run_suite(args: SuiteArgs) -> anyhow::Result<()> {
         );
         return Ok(());
     }
-    render(&args.bench_dir)
-}
-
-/// Rewrites the engine table in `RESULTS.md` from the receipts on disk,
-/// leaving the header and Open notes outside the markers untouched.
-const BEGIN: &str = "<!-- BEGIN ENGINE TABLE (generated by `ferrox bench --render`) -->";
-const END: &str = "<!-- END ENGINE TABLE -->";
-
-pub fn render(bench_dir: &Path) -> anyhow::Result<()> {
-    let dir = engine_receipt_dir(bench_dir);
-    let mut receipts: Vec<serde_json::Value> = Vec::new();
-    if dir.is_dir() {
-        let mut paths: Vec<PathBuf> = std::fs::read_dir(&dir)?
-            .filter_map(|e| e.ok().map(|e| e.path()))
-            .filter(|p| p.extension().is_some_and(|e| e == "json"))
-            .collect();
-        paths.sort();
-        for p in paths {
-            if let Ok(text) = std::fs::read_to_string(&p) {
-                if let Ok(v) = serde_json::from_str::<serde_json::Value>(&text) {
-                    receipts.push(v);
-                }
-            }
-        }
-    }
-
-    // No receipts at all means there is nothing to render. Writing the
-    // table anyway would replace a real ledger with an empty one and
-    // report success, which is worse than doing nothing.
-    if receipts.is_empty() {
-        anyhow::bail!(
-            "no engine receipts under {}, so there is nothing to render. \
-             Run `ferrox bench --suite` first.",
-            dir.display()
-        );
-    }
-
-    let suite = load_suite(bench_dir).unwrap_or_default();
-    let name_of = |id: &str| {
-        suite
-            .iter()
-            .find(|e| e.id == id)
-            .map(|e| e.name.clone())
-            .unwrap_or_else(|| id.to_string())
-    };
-
-    #[derive(Clone)]
-    struct Row {
-        /// Which machine produced this row. Rows never merge across
-        /// hosts; see the grouping below.
-        host: String,
-        model: String,
-        backend: String,
-        test: String,
-        ferrox: Option<f64>,
-        llama: Option<f64>,
-        gap: Option<f64>,
-    }
-
-    // Rows from two machines are not one table.
-    //
-    // `render` reads every receipt in the directory, so the moment a
-    // second host writes one, its numbers would sort in beside this
-    // one's under a single heading with nothing saying so. A reader
-    // comparing a 5.06x row against a 1.41x row would be comparing two
-    // computers. Receipts written before 0.13.0 carry no spec at all,
-    // and they group together as one unknown host rather than being
-    // waved through individually.
-    let mut hosts: std::collections::BTreeSet<String> = std::collections::BTreeSet::new();
-    for r in &receipts {
-        hosts.insert(host_identity(r));
-    }
-    // More than one host used to be a hard refusal, on the grounds
-    // that "one table cannot describe them". That reasoning is right
-    // and is kept: a gap is only meaningful against the machine it was
-    // measured on. What changed is the remedy. Refusing meant the
-    // ledger could only ever describe whichever laptop happened to run
-    // the suite, which is how it came to claim a CPU gap of 1.41x to
-    // 5.06x while saying nothing about x86 or CUDA at all. Rows are now
-    // grouped into one section per host, so two machines coexist
-    // without any row being compared against the wrong one.
-
-    let mut rows: Vec<Row> = Vec::new();
-    for r in &receipts {
-        let host_label = host_identity(r);
-        let id = r.get("id").and_then(|v| v.as_str()).unwrap_or("?");
-        let backend = r
-            .get("backend")
-            .and_then(|v| v.as_str())
-            .unwrap_or("?")
-            .to_string();
-        let Some(tests) = r.get("tests").and_then(|v| v.as_array()) else {
-            continue;
-        };
-        for t in tests {
-            rows.push(Row {
-                host: host_label.clone(),
-                model: name_of(id),
-                backend: backend.clone(),
-                test: t
-                    .get("test")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("?")
-                    .to_string(),
-                ferrox: t.get("ferrox_tps").and_then(|v| v.as_f64()),
-                llama: t.get("llama_tps").and_then(|v| v.as_f64()),
-                gap: t.get("gap").and_then(|v| v.as_f64()),
-            });
-        }
-    }
-
-    // Worst-first within (backend, test): high gap = ferrox farther behind.
-    rows.sort_by(|a, b| {
-        let backend_ord = |s: &str| match s {
-            "metal" => 0,
-            "cuda" => 1,
-            "cpu" => 2,
-            _ => 3,
-        };
-        let test_ord = |s: &str| {
-            if s.starts_with("pp") {
-                0
-            } else if s.starts_with("tg") {
-                1
-            } else {
-                2
-            }
-        };
-        a.host
-            .cmp(&b.host)
-            .then_with(|| backend_ord(&a.backend).cmp(&backend_ord(&b.backend)))
-            .then_with(|| test_ord(&a.test).cmp(&test_ord(&b.test)))
-            .then_with(|| {
-                b.gap
-                    .partial_cmp(&a.gap)
-                    .unwrap_or(std::cmp::Ordering::Equal)
-            })
-            .then_with(|| a.model.cmp(&b.model))
-    });
-
-    let mut table = String::new();
-    table.push_str(BEGIN);
-    table.push_str("\n\n## Engine (`ferrox bench` vs `llama-bench`)\n\n");
-    // Which machine, stated in the generated block rather than in prose
-    // above it, so it cannot drift away from the numbers it describes.
-    {
-        let n = hosts.len();
-        if n == 1 {
-            table.push_str(&format!(
-                "Measured on: **{}**\n\n",
-                hosts.iter().next().cloned().unwrap_or_default()
-            ));
-        } else if n > 1 {
-            table.push_str(&format!(
-                "Measured on **{n} hosts**, one section each. Rows are never \
-                 compared across machines.\n\n"
-            ));
-        }
-    }
-    // A summary table, not prose. Generated from the same receipts as
-    // the detail rows below, so it cannot drift away from them the way
-    // a hand-written headline does. One line per host and backend: the
-    // range is what a reader wants before any individual model.
-    {
-        use std::collections::BTreeMap;
-        let mut by: BTreeMap<(String, String, bool), Vec<f64>> = BTreeMap::new();
-        for r in &rows {
-            if let Some(g) = r.gap {
-                by.entry((r.host.clone(), r.backend.clone(), r.test.starts_with("pp")))
-                    .or_default()
-                    .push(g);
-            }
-        }
-        if !by.is_empty() {
-            table.push_str("### Summary\n\n");
-            table.push_str("| Host | Backend | Prefill gap | Decode gap |\n");
-            table.push_str("|---|---|---|---|\n");
-            let mut seen: Vec<(String, String)> =
-                by.keys().map(|(h, b, _)| (h.clone(), b.clone())).collect();
-            seen.dedup();
-            for (host, backend) in seen {
-                let fmt = |pp: bool| -> String {
-                    match by.get(&(host.clone(), backend.clone(), pp)) {
-                        Some(v) if !v.is_empty() => {
-                            let lo = v.iter().cloned().fold(f64::INFINITY, f64::min);
-                            let hi = v.iter().cloned().fold(f64::NEG_INFINITY, f64::max);
-                            if (hi - lo).abs() < 0.005 {
-                                gap_cell(lo)
-                            } else {
-                                format!("{} to {}", gap_cell(lo), gap_cell(hi))
-                            }
-                        }
-                        _ => "—".to_string(),
-                    }
-                };
-                table.push_str(&format!(
-                    "| {host} | {} | {} | {} |\n",
-                    backend.to_uppercase(),
-                    fmt(true),
-                    fmt(false)
-                ));
-            }
-            table.push('\n');
-        }
-    }
-
-    fn push_section_at(table: &mut String, depth: &str, title: &str, rows: &[&Row]) {
-        if rows.is_empty() {
-            return;
-        }
-        table.push_str(&format!("{depth} {title}\n\n"));
-        table.push_str("| Model | Test | ferrox tok/s | llama.cpp tok/s | Gap |\n");
-        table.push_str("|---|---|---|---|---|\n");
-        for r in rows {
-            table.push_str(&format!(
-                "| {} | {} | {} | {} | {} |\n",
-                r.model,
-                r.test,
-                r.ferrox
-                    .map(|v| format!("**{v:.2}**"))
-                    .unwrap_or_else(|| "—".into()),
-                r.llama
-                    .map(|v| format!("**{v:.2}**"))
-                    .unwrap_or_else(|| "—".into()),
-                r.gap.map(gap_cell).unwrap_or_else(|| "—".into()),
-            ));
-        }
-        table.push('\n');
-    }
-
-    let metal: Vec<&Row> = rows.iter().filter(|r| r.backend == "metal").collect();
-    let cuda: Vec<&Row> = rows.iter().filter(|r| r.backend == "cuda").collect();
-    let cpu: Vec<&Row> = rows.iter().filter(|r| r.backend == "cpu").collect();
-    let other: Vec<&Row> = rows
-        .iter()
-        .filter(|r| !matches!(r.backend.as_str(), "metal" | "cuda" | "cpu"))
-        .collect();
-
-    if rows.is_empty() {
-        table.push_str("| _no engine receipts yet_ | | | | | |\n\n");
-    } else if hosts.len() <= 1 {
-        push_section_at(&mut table, "###", "Metal", &metal);
-        push_section_at(&mut table, "###", "CUDA", &cuda);
-        push_section_at(&mut table, "###", "CPU", &cpu);
-        push_section_at(&mut table, "###", "Other backends", &other);
-    } else {
-        // One section per machine. A reader scanning for a gap sees the
-        // host before the number, which is the only order in which the
-        // number means anything.
-        for host in &hosts {
-            let here: Vec<&Row> = rows.iter().filter(|r| &r.host == host).collect();
-            table.push_str(&format!("### {host}\n\n"));
-            for (title, backend) in [("Metal", "metal"), ("CUDA", "cuda"), ("CPU", "cpu")] {
-                let sub: Vec<&Row> = here
-                    .iter()
-                    .copied()
-                    .filter(|r| r.backend == backend)
-                    .collect();
-                push_section_at(&mut table, "####", title, &sub);
-            }
-            let sub: Vec<&Row> = here
-                .iter()
-                .copied()
-                .filter(|r| !matches!(r.backend.as_str(), "metal" | "cuda" | "cpu"))
-                .collect();
-            push_section_at(&mut table, "####", "Other backends", &sub);
-        }
-    }
-
-    table.push_str(END);
-
-    let results = bench_dir.join("RESULTS.md");
-    let existing = std::fs::read_to_string(&results).unwrap_or_default();
-    let updated = splice(&existing, &table);
-    std::fs::write(&results, updated)?;
-    eprintln!(
-        "ferrox bench: engine table written to {}",
-        results.display()
-    );
-    Ok(())
-}
-
-fn gap_cell(g: f64) -> String {
-    let marker = if g < 0.95 {
-        "🟢"
-    } else if g <= 1.05 {
-        "⚪"
-    } else {
-        "🔴"
-    };
-    format!("{marker} **{g:.2}×**")
-}
-
-/// Replaces the marked block, or appends it if the markers are absent.
-fn splice(existing: &str, block: &str) -> String {
-    if let (Some(start), Some(end)) = (existing.find(BEGIN), existing.find(END)) {
-        let mut out = String::with_capacity(existing.len() + block.len());
-        out.push_str(&existing[..start]);
-        out.push_str(block);
-        out.push_str(&existing[end + END.len()..]);
-        return out;
-    }
-    let mut out = existing.to_string();
-    if !out.ends_with('\n') {
-        out.push('\n');
-    }
-    out.push('\n');
-    out.push_str(block);
-    out.push('\n');
-    out
+    crate::bench_render::render(&args.bench_dir)
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    /// Writes one engine receipt for `host` into `dir`.
-    fn receipt(dir: &Path, id: &str, host: &str, backend: &str, ferrox: f64, llama: f64) {
-        let r = serde_json::json!({
-            "schema": 2, "kind": "engine", "id": id,
-            "backend": backend, "backend_active": backend,
-            "host_spec": {"label": host},
-            "tests": [{"test": "tg128", "ferrox_tps": ferrox, "llama_tps": llama,
-                       "gap": llama / ferrox}],
-        });
-        std::fs::write(
-            dir.join(format!("{id}_{backend}.json")),
-            serde_json::to_string(&r).expect("json"),
-        )
-        .expect("write receipt");
-    }
-
-    /// Two machines render as two sections, and neither is dropped.
-    ///
-    /// This used to be a hard refusal ("one table cannot describe
-    /// them"), which meant the ledger could only ever describe the
-    /// laptop that happened to run the suite: it claimed a CPU gap of
-    /// 1.41x to 5.06x while having no x86 or CUDA row at all. The
-    /// refusal's REASON was right, so rows are separated rather than
-    /// merged, and this pins that.
-    #[test]
-    fn two_hosts_render_as_two_sections_rather_than_an_error() {
-        let dir = std::env::temp_dir().join(format!(
-            "ferrox_render_{}_{:?}",
-            std::process::id(),
-            std::thread::current().id()
-        ));
-        let engine = dir.join("receipts").join("engine");
-        std::fs::create_dir_all(&engine).expect("mkdir");
-        std::fs::write(
-            dir.join("RESULTS.md"),
-            format!("head\n{BEGIN}\nold\n{END}\ntail\n"),
-        )
-        .expect("seed");
-        receipt(&engine, "m1", "Apple M2 Pro", "metal", 100.0, 90.0);
-        receipt(&engine, "m1", "Rented Xeon", "cpu", 10.0, 20.0);
-
-        render(&dir).expect("two hosts must render, not refuse");
-
-        let out = std::fs::read_to_string(dir.join("RESULTS.md")).expect("read");
-        assert!(out.contains("Apple M2 Pro"), "first host missing:\n{out}");
-        assert!(out.contains("Rented Xeon"), "second host missing:\n{out}");
-        assert!(
-            out.contains("2 hosts"),
-            "the reader is not told there are two machines:\n{out}"
-        );
-        // The whole point: a row is never presented without its host.
-        let xeon = out.find("Rented Xeon").expect("host heading");
-        let cpu_row = out.find("| 10.00 |").or_else(|| out.find("**10.00**"));
-        if let Some(cpu_row) = cpu_row {
-            assert!(
-                cpu_row > xeon,
-                "the Xeon's row appears before its host heading"
-            );
-        }
-        let _ = std::fs::remove_dir_all(&dir);
-    }
 
     /// Two machines running the same entry must not share a filename.
     #[test]
@@ -687,107 +316,6 @@ mod tests {
             "two hosts sharing a receipt name is how a ledger loses a machine"
         );
         assert_eq!(host_slug("  --  "), "");
-    }
-
-    /// The summary is GENERATED, not written by hand.
-    ///
-    /// It replaced a hand-written headline table, which is a thing that
-    /// drifts: the numbers above the fold stop matching the receipts
-    /// below it and nobody notices, because nothing compares them. One
-    /// row per host and backend, with the range taken from the same
-    /// rows the detail tables use.
-    #[test]
-    fn the_summary_is_derived_from_the_same_rows_as_the_detail_tables() {
-        let dir = std::env::temp_dir().join(format!(
-            "ferrox_summary_{}_{:?}",
-            std::process::id(),
-            std::thread::current().id()
-        ));
-        let engine = dir.join("receipts").join("engine");
-        std::fs::create_dir_all(&engine).expect("mkdir");
-        std::fs::write(
-            dir.join("RESULTS.md"),
-            format!("head\n{BEGIN}\nold\n{END}\ntail\n"),
-        )
-        .expect("seed");
-        // Two gaps on one host+backend, so the summary must show a range.
-        receipt(&engine, "a", "Box One", "cuda", 10.0, 20.0);
-        receipt(&engine, "b", "Box One", "cuda", 10.0, 100.0);
-
-        render(&dir).expect("render");
-        let out = std::fs::read_to_string(dir.join("RESULTS.md")).expect("read");
-
-        assert!(out.contains("### Summary"), "no summary table:\n{out}");
-        let summary = &out[out.find("### Summary").expect("summary")..];
-        let first_detail = summary.find("\n### ").unwrap_or(summary.len());
-        let summary = &summary[..first_detail];
-        assert!(
-            summary.contains("2.00×") && summary.contains("10.00×"),
-            "the summary must span the rows it describes:\n{summary}"
-        );
-        assert!(
-            summary.contains("Box One"),
-            "the summary must name the host:\n{summary}"
-        );
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn gap_cell_colours_match_the_ledger_convention() {
-        assert!(gap_cell(0.80).starts_with("🟢"));
-        assert!(gap_cell(1.00).starts_with("⚪"));
-        assert!(gap_cell(0.96).starts_with("⚪"));
-        assert!(gap_cell(1.40).starts_with("🔴"));
-    }
-
-    #[test]
-    fn splice_replaces_an_existing_block_and_keeps_the_surrounding_text() {
-        let doc = format!("before\n{BEGIN}\nold\n{END}\nafter\n");
-        let out = splice(&doc, &format!("{BEGIN}\nnew\n{END}"));
-        assert!(out.contains("before"), "text before the block must survive");
-        assert!(out.contains("after"), "text after the block must survive");
-        assert!(out.contains("new"));
-        assert!(!out.contains("old"), "the old block must be gone");
-    }
-
-    #[test]
-    fn splice_appends_when_the_markers_are_missing() {
-        let out = splice("just some prose\n", &format!("{BEGIN}\nfresh\n{END}"));
-        assert!(out.starts_with("just some prose"));
-        assert!(out.contains("fresh"));
-    }
-
-    /// A render with no receipts must refuse rather than publish an
-    /// empty table over a real one.
-    #[test]
-    fn rendering_nothing_refuses_instead_of_emptying_the_ledger() {
-        let dir = std::env::temp_dir().join(format!("ferrox-render-guard-{}", std::process::id()));
-        let _ = std::fs::remove_dir_all(&dir);
-        std::fs::create_dir_all(dir.join("receipts").join("engine")).unwrap();
-        let results = dir.join("RESULTS.md");
-        let original = "# Results\n\nreal numbers live here\n";
-        std::fs::write(&results, original).unwrap();
-
-        let err = render(&dir).unwrap_err().to_string();
-        assert!(
-            err.contains("nothing to render"),
-            "expected a refusal naming the empty receipt dir, got: {err}"
-        );
-        assert_eq!(
-            std::fs::read_to_string(&results).unwrap(),
-            original,
-            "the existing ledger must survive a render that had no receipts"
-        );
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn splice_does_not_duplicate_the_block_on_a_second_render() {
-        let block = format!("{BEGIN}\nv1\n{END}");
-        let once = splice("doc\n", &block);
-        let twice = splice(&once, &format!("{BEGIN}\nv2\n{END}"));
-        assert_eq!(twice.matches(BEGIN).count(), 1, "exactly one engine block");
-        assert!(twice.contains("v2") && !twice.contains("v1"));
     }
 }
 
@@ -852,6 +380,7 @@ mod committed_receipt_tests {
 mod host_identity_tests {
     use super::host_identity;
 
+    /// A receipt carrying just the fields `host_identity` reads.
     fn receipt(label: &str, accelerator: Option<&str>) -> serde_json::Value {
         let mut r = serde_json::json!({"host_spec": {"label": label}});
         if let Some(card) = accelerator {
